Workaround to avoid port conflict
Usually on the modern installations of linux, there is a chance that systemd-resolved.service is running. The suggested way to install usign docker on any such machine is to follow this guide https://github.com/pi-hole/docker-pi-hole#installing-on-ubuntu but it may be a little complex.

To resolve this, I assign a static IP to my physical device that hosts the pi-hole and I mention the static IP in the docker-compose file for pi-hole in the form of dns: property and supplying the IP address for the ports as well