Fix issue with auto restart after power outage
In case of a power outage when the router and the pi-hole both stopped, the pi-hole was not restarting on its own and needed manual intervention. The root cause of this issue was the fact that I was previously using an IP address in the ports section.

Now what happens is that while the router is starting, the host of the pi-hole is not assigned an IP address until the DHCP services on the router are started. This would mean that the IP address of the host is unknown but the docker container wants to bind on the provided IP address and port.

To resolve this issue, the host needs to stop using the systemd-resolved stub and start to use the actual file and a DNS should be assigned using netplan. Optionally, a static IP can also be assigned to the host using the netplan configuration.